Refactor frontend and implement websockets
I refactored the frontend logic of various pages and components like the
audio AudioPlayer and added some functionalities that simplify data
fetching as well user context saving and retrieval operations.
I also implemented the websockets connection on the Checkpoints page
when the user is logged in to make the system perform better
in real time.